更正表格資料型態

+++ database/migrations/2024_10_22_031251_create_agency_statistics_table.php
@@ -21,8 +21,8 @@ class CreateAgencyStatisticsTable extends Migration
             $table->integer('simple_appointees');
             $table->integer('recommended_appointees');
             $table->integer('commissioned_appointees');
-            $table->decimal('average_age', 4, 2);
-            $table->decimal('average_seniority', 4, 2);
+            $table->double('average_age', 4, 2);
+            $table->double('average_seniority', 4, 2);
             $table->timestamps();
         });
     }
